+++ HISTORY.md
@@ -5,6 +5,7 @@ unreleased
     - i.e. `'path/to/json'` will no longer return `application/json`
   * Improve generation of extension to MIME mapping
   * Refactor internals for readability and no argument reassignment
+  * Prefer `application/*` MIME types from the same source
 
 2.0.14 / 2015-06-06
 ===================